@@ -28,10 +28,9 @@ readme = "README.md"
 #     "LICENSE",
 #     "THIRD_PARTY_NOTICES.md",
 # ]
-requires-python = ">=3.8"  # python_requires is also located in setup.py
+requires-python = ">=3.9"  # python_requires is also located in setup.py
 classifiers = [
     "Development Status :: 5 - Production/Stable",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -105,7 +104,6 @@ git_describe_command = 'git describe --dirty --tags --long --match "*.*.*"'
 [tool.ruff]
 output-format = "grouped"
 line-length = 120
-target-version = "py38"
 force-exclude = true  # Fixes issue with megalinter config preventing exclusion of files
 extend-exclude = [
     "newrelic/packages/",
