@@ -27,17 +27,12 @@ from newrelic.common.package_version_utils import (
 )
 
 # Notes:
-# importlib.metadata was a provisional addition to the std library in PY38 and PY39
+# importlib.metadata was a provisional addition to the std library in Python 3.8 and 3.9
 # while pkg_resources was deprecated.
-# importlib.metadata is no longer provisional in PY310+.  It added some attributes
+# importlib.metadata is no longer provisional in Python 3.10+. It added some attributes
 # such as distribution_packages and removed pkg_resources.
 
-IS_PY38_PLUS = sys.version_info[:2] >= (3, 8)
 IS_PY310_PLUS = sys.version_info[:2] >= (3, 10)
-SKIP_IF_NOT_IMPORTLIB_METADATA = pytest.mark.skipif(not IS_PY38_PLUS, reason="importlib.metadata is not supported.")
-SKIP_IF_IMPORTLIB_METADATA = pytest.mark.skipif(
-    IS_PY38_PLUS, reason="importlib.metadata is preferred over pkg_resources."
-)
 SKIP_IF_NOT_PY310_PLUS = pytest.mark.skipif(not IS_PY310_PLUS, reason="These features were added in 3.10+")
 
 
@@ -101,7 +96,6 @@ def test_get_package_version_tuple(monkeypatch, attr, value, expected_value):
     assert version == expected_value
 
 
-@SKIP_IF_NOT_IMPORTLIB_METADATA
 @validate_function_called("importlib.metadata", "version")
 def test_importlib_dot_metadata():
     # Test for importlib.metadata from the standard library.
@@ -109,26 +103,9 @@ def test_importlib_dot_metadata():
     assert version not in NULL_VERSIONS, version
 
 
-@SKIP_IF_IMPORTLIB_METADATA
-@validate_function_called("importlib_metadata", "version")
-def test_importlib_underscore_metadata():
-    # Test for importlib_metadata, a backport library available on PyPI.
-    version = get_package_version("pytest")
-    assert version not in NULL_VERSIONS, version
-
-
 @SKIP_IF_NOT_PY310_PLUS
 @validate_function_called("importlib.metadata", "packages_distributions")
 def test_mapping_import_to_distribution_packages():
-    version = get_package_version("pytest")
-    assert version not in NULL_VERSIONS, version
-
-
-@SKIP_IF_IMPORTLIB_METADATA
-@validate_function_called("pkg_resources", "get_distribution")
-def test_pkg_resources_metadata(monkeypatch):
-    # Prevent importlib_metadata from being used by these tests
-    monkeypatch.setitem(sys.modules, "importlib_metadata", None)
     version = get_package_version("pytest")
     assert version not in NULL_VERSIONS, version
 
